fix: RoPE 在 bf16 autocast 下报错（GPU 模式）+ 声明 tokenizers 依赖
GPU 模式下训练用 torch.autocast(bf16)，而 view_as_complex 不支持 bf16，
导致 apply_rotary_pos_emb 在预训练/SFT/DPO 前向时崩溃（CPU 验证不触发）。
修复：旋转前先 .float()，末尾 .type_as() 还原 dtype（与 RMSNorm 同文件同款写法）。
同步修复 02/03/05 三处相同实现。

远程 GPU 全流程验证通过：01→08 脚本 + 作业测试（骨架 0/7 skip、参考实现 7/7）。

+++ courses/Part7_minimind/scripts/02_rmsnorm_rope.py
@@ -128,8 +128,9 @@ def apply_rotary_pos_emb(q, k, freqs_cis):
     half = head_dim // 2
     freqs_cis = freqs_cis[:T].view(T, 1, half)  # (T, 1, half) 广播到各 head
     # (B, T, H, half, 2) → 复数 (B, T, H, half)
-    q_ = torch.view_as_complex(q.reshape(B, T, num_heads, half, 2))
-    k_ = torch.view_as_complex(k.reshape(B, T, num_heads, half, 2))
+    # view_as_complex 不支持 bf16（autocast 下会报错），先转 fp32，末尾 type_as 还原
+    q_ = torch.view_as_complex(q.float().reshape(B, T, num_heads, half, 2))
+    k_ = torch.view_as_complex(k.float().reshape(B, T, num_heads, half, 2))
     q_out = torch.view_as_real(q_ * freqs_cis).reshape(B, T, num_heads, head_dim)
     k_out = torch.view_as_real(k_ * freqs_cis).reshape(B, T, num_heads, head_dim)
     return q_out.type_as(q), k_out.type_as(k)
